Implemented Melee Combat Framework
Created; a struct and datatable for melee weapons, an interface for the weapons, a light attack component

The component isn't attatched to the player yet, but it has the basics ready. It needs animations which I need to pull now. The struct currently only has information for melee weapons (if that).