More spell projectile construction script and more components on spell projectile blueprint. Base spell niagara system created with fountain template.
Some more functionality in the spell projectile just setting the static mesh based on spell shape but need to actually put the right meshes in and also setting the initial speed of the projectile component that has also been added to the projectile blueprint.

Created a base spell niagara system but was going to use a mesh render for the main shape of the spell but can't change it based on user parameter so not too sure about how to do it but currently this system is just a fountain template system with velocity added from point instead of in cone.