Reorganised some things in the combat section. More setup for combat.
Did a little reorganising of the combat section of the folders.

Created a new struct for general weapon info containing the name (which is a newly created enum for all weapon names) as well as whether it is ranged as a bool.

New enums: weapon names, melee weapon names, ranged weapon names.

New struct for ranged weapon info.

New component for spells that will be on characters that can use spells allowing them to switch between spells etc and get the currently equipped spell (think of it as an armoury component).

New interface for characters that have weapons: BPI_Has_Weapon which will allow for easy retrieval of the currently equipped weapon independent of if it is ranged or not which will allow for easy retrieval of the general info of the weapon to determine its name and if it is ranged.