Spells update. Ranged combat update and attack handler update. Input Update. Added interfaces to player.
Spells component added functionality basic for switching current spell, will need work on what spells are given to each character but that will come with time and from the character info data table once the character data struct is created.
Will also need to add UI function calling from this component to actually show the change to the current spell.

More work on the ranged attack via the attack handler using the basic attack input action that will work for both melee and ranged combat.

Also added 2 bools to the ranged attack component: charging spell and firing spell which will be used for checking to see if the attack can be done in the attack handler among other checks.

Inputs have been foldered now in the input section of the player to make clearer the intention behind each input. Also added in controller inputs for the inputs that were already there and for the new inputs which are next spell and previous spell which are used within the spells component.