Character Info Struct. Fixed setting the weapon owner in the attach weapon component. AOE component. Split component. Homing component. Spell on hit function.
Structure containing all necessary info for a character including players and enemies so this struct is replacing the enemies struct as well since they will have the same info.

In the attach weapon component just added getting the owner of the component and then setting this to be the owner of the weapon which should fix Ben's issues.

AOE component which has the AOE function within it to do all necessary functionality surrounding it. Starts with a sphere trace and then goes from there using the spell's AOE data to do damage and knockback. Also setup for doing persistent AOE functionality in this component but also need to actually create the persistent AOE blueprint before this.

Split component that has the Split function within it to spawn more of the same spell projectiles without the split characteristic and also with the lowered stats that are set through the split info in the spell info struct for the shot spell projectile.

Homing component that has 2 functions in it which are Activate homing and deactivate homing which do as they say using the owner's projectile movement component. Will also need to do some more stuff in this component to set whether the projectile should be homing based on proximity to characters based on the homing info in the spell info. Will also need to do stuff with homing strength in this component as well.

Spell on hit function using the built in projectile movement component on stop function to then do a sphere trace based on the scale of the spell which will detect pawns and physics bodies to apply damage as necessary and also some knockback based on velocity (temp since this will definitely be too much). Also this will activate the AOE function and split function if the spell has that component, since the components are added in begin play on the projectile based on its characteristics.