Fixed weapon collision on attaching to character. Gave static mesh to weapon placeholder in data table. Changed set weapon info in BPI_Melee to use weapon names enum.
Set actor enable collision false after attaching the weapon to the character in the attach weapon component.

Set katana mesh as the static mesh for the placeholder weapon in the melee weapon data table.

BPI_Melee set weapon info now uses the enum weapon names instead of a name to limit errors with case sensitivity.
Weapon parent now has its name as a enum of weapon names instead of a name to limit errors with case sensitivitiy.
This enum is now also instance editable and editable on spawn so can change it when spawning the weapon in attach weapon component.