Ranged weapon data table. General info struct for weapons now contains the static mesh of the weapon as well. Weapon parent blueprint set info update.
Ranged weapon data table that will contain any specific information pertaining to the ranged weapons similar to the melee weapon data table.

General weapon info struct now contains the static mesh of the weapon since this is a shared factor between ranged and melee weapons and so it can be contained within here and then within the weapons struct/data table.

The weapon blueprint set info function now uses the weapon data table to get the information instead of just the melee data table so that it should work for both melee and ranged weapons.