Changed staff build settings and pivot location and rotation. Attach weapon component update.
Staff now has smaller build settings and better pivot for being attached to the hand.

Attach weapon component now has the weapon name for the weapon being a public and editable on spawn enum variable of type weapon names so can change it in the player and other places through get component for the attach weapon component.