Setup basic and most likely temporary functionality for setting ranged attacking and melee attacking and weapon name in animation blueprint. Weapon interface. Weapon interface functions.
Functionality for setting ranged and melee attacking bool variables and updating them with blueprint update in animation blueprint depending on the light attacks component as of now and the ranged attack component.

Also getting the equipped weapon and getting its name to then store as a enum variable in the animation blueprint.

Removed the set weapon info function from the melee interface and instead added it to the new weapon interface since this is independent on whether the weapon is melee or ranged.

New interface generally for weapons so not melee or ranged specific functions.