Animation blueprint update. Can attack in attack handler.
Isvalid checks for the components.

Can check combat info with very small delay just to get rid of some errors.

Can attack in attack handler currently set by not the attacking in the light attacks and not firing a spell in the ranged attack component.

Theoretically the animation blueprint is now working with the correct blending just need to add the content since I tested it with a ranged weapon and it did a blank animation as it should.