Ranged weapon struct update. Spells update. Animation update.
Staff static mesh fixed.

Ranged weapon struct now has the spells within it that are then set in the spells component with the current spell just set as the first of the array.

Spells construction script now also done in the begin play.

Added first state in ranged weapon state machine that works correctly and as intended just need to get working why the projectile is not able to be seen etc.