Animation changes and weapon changes
implemented take damage (cant check if works yet bc struggling to get animations to work)

Re-done the animations and animation notifiers for the example katana melee combo.

Created a new ranged and melee weapon instead of the variable in the player

Spawn in another weapon in the attach weapon component.

Created states for the attacking animations in the AnimBP.

Created Queued attacks

The logic is basically there for it to work but something aint clicking and im gonna look into that next time i get on.