Spells switching fixed. Burst Angling Fixed. Default values in all spell characteristics that were needed for now. For loops fixed. Split fixed. AOE basics work needs more functionality tho. Homing needs actual homing target detection.
Spells component switching now properly works with the current spell index and the looping back around to the last or first index of the spells array works now.

Burst angling and general way to calculate the angles for characteristics that have radius' since it was not working properly.

Default values have now been setup for all the spell characteristics with the majority of the fields filled in and only a couple have not been filled just to make editing them easier since we have a baseline now, some of the values are not right since I don't know the correct value.

For loops are now fixed for the spells since they were looping one more time than necessary since they are inclusive and I forgot.

Split characteristic is now fixed to split the spell into the correct amount of spells with the correct angling and doing the end of life of the original spell and also setting the owner of the split spells correctly now as well.

AOE basics works from testing once but will need more functionality like knockback etc.

Homing needs actual target delegation since rn it doesn't do that and so doesn't actually home in on enemies.