Projectile EOL Fix. Health and Death update.
We now have natural end of life and non-natural end of life for spells since when a spell naturally dies after its duration it was doing hit detection at 0,0,0 instead of where it currently was so this is now fixed.

Health update.
Bloodsplosion is now implemented in the death function within the health component.

Death component deleted.

Slight camera shake on take damage.
Blood effect added to enemies.
Blood effect call and such in take damage function.
Sound attenuations have larger ranges since the camera is much further away from the sound location.

Destroy character and associated weapons function in health component that does as it says and is called in the death function after a very small delay.

Knockback on damage is now consolidated in the take damage function in the health component using 2 parameters knockback strength horizontal and knockback strength vertical.