Spell Cooldown system yayayayayayay NO TUTORIAL NEEDED LETS FUCKIN GO
Spell cooldown system contained within the spells component that uses the component tick and a function.

The function is what starts the cooldown and is called in the shoot event in the ranged attack component with the corresponding spell info.

New struct of spell cooldown containing 3 variables: float for the time the cooldown started, float for the cooldown duration and a int for the index of the spell in the spells array in the spells component.

Uses 3 new variables in the spells component: spells cooldowns an array of type spell cooldown struct, an array called spell cooldowns active which is an array of bools, and an array of integers called spell cooldowns to remove.

Spell cooldown system could be more optimised but since the characters using spells will not have tens of spells and instead only a couple it shouldn't be too inefficient.

Also in the attack handler now we check to see if the spell we are trying to use is on cooldown and if it is then the spell isn't fired however if it isn't then we successfully shoot the spell.