Spell collision update. Camera update. AOE damage and knockback falloff fix. Melee simplification and "fix"
Spell collision type is now custom with block all as a template and then ignoring destruction collision types which is what the spells now are so that they ignore each other. May in the future change this so there is a small object within the spell that other objects can hit but not too sure.

Made the camera spring arm shorter so more zoomed in for now can always change this easily in the camera blueprint.

Actually got the AOE damage falloff and knockback falloff working since the calculation for the percentage for this dependent on the location of the hit character and the location of the epicentre wasn't right even though I thought it was.

"Fixed" the melee attacking. Simplified it to just attacking instead of attacking animation as bools but kept them in for now. The melee attacks work for comboing however the animations are a little janky with some jittering but the main thing is that they are working and you can continuously combo attacks with no issues, I think. Can alway tweak it and add more anyway and if its not how you want then change it for sure.

Also the attacks start on the second animation, just the way it works, you can try to change it if you want idk.