Movement Update. Movement Speed Manager. Character struct update. Renamed Struct_MeleeWeapons and associated data table.
New component called BPC_Movement_Manager which will manage the movement speeds of the character it is attached to. It contains 2 variables as of now: the basic movement speed of the owning character and the movement speed multiplier. Then in the event tick we are constantly finterping to a target move speed which is the base move speed multiplied by the multiplier which will be used in various aspects of the game to change how fast a character can move. For example when a character is charging up a spell their movement speed is decreased and so by changing this multiplier we can slow the character down smoothly over a very short period of time and then allow them to move back at their normal movement speed afterwards very easily.

New interface for general characters. This has a function within it to get the character info which is used in the movement manager to get the basic movement info from the character movement struct. This interface has been added to test dummy enemy and player blueprint as well.

Renamed the meleeweapons struct to be Struct_Melee_Weapon_Info to be more inline with the naming of other structs. Character info now has 2 structs within it for specifically the ranged weapon and the melee weapon of the character. Just need to setup for the attach weapon component to use this info from the character struct.

Also added a character struct variable into the player blueprint and the test dummy blueprint instead of using a specific health info struct since the health info struct is contained within the character info struct.