Spell cooldown re-implementation in attack handler. Charge characteristics movement speed slow. Movement speed multiplier updated/changed. Shoot screen shake. Burst randomisation of speed and scale. Spell cooldown fix. Split opposite direction functionality implemented. Split stat percentages in children work
Spell cooldowns checking is now re-implemented in the attack handler when attempting to shoot a spell since I think it was overwritten in another commit.

Now when charging a spell the movement speed multiplier is updated accordingly depending on the percentage slow in the charge info of the spell and then is undone when the charging stops or the spell is fired.

Movement speed multiplier is now set in a specific function that will subtract the passed decimal from the current speed multiplier so to add speed back again you must pass a negative float. Which easily allows for multiple changes to the movement speed to happen at the same time and as long as the opposite calculation is done to the movement speed multiplier then it will eventually go back to the normal speed multiplier.

New blueprint for spell shooting screenshake that is used in the ranged attack component in the specific shoot camera shake function which scales the camera shake based on the damage of the spell for now which could change in the future. This scale is clamped however so it isn't overly egregious. The screenshake isn't played in the world for this and instead is just played generally for the "player character 1" camera manager which should be fine and should be played no matter the distance since knowing about spells is pretty important for the players to know.

Burst characteristic randomisation of speed and scale of the spells is now fully implemented in the shoot notifiy in the player animation blueprint. The float values in the corresponding burst struct is the range so 0.5 would be 0.25 lower and 0.25 higher than the original for the stats.

Shoot function in the ranged attack component now takes another spell struct for the original spell which is used in cases such as burst (with randomisation active) so that the start spell cooldown function will take the original spell info to check against its spells array. Since this was causing issues where the passed spell info and the spell info in the array couldn't be found and the spell wouldn't go on cooldown.

Split characteristic splitting in opposite direction functionality is now fully implemented in the split component.

Split stat percentage changes in children tested and working as it should.