AOE damage and knockback falloff maximums should be implemented - kinda hard to test.
AOE damage and knockback falloff maximums implemented with clamps but its kinda hard to test but I think it is working as intended.

AOE persistent blueprint and functionality. This spawns a decal which is technically the blood puddle decal with its colour set right now to blue but can be set by parameter in the future. This blueprint has 3 passed variables to the blueprint: duration, percentage slow, and radius. It also has an array of actors which is used to keep track of the actors affected by its slow so it doesn't repeatedly do it.

The blueprint only contains a sphere collision which is set to only overlap pawns and ignore every other collision type and to not allow the player to stand on it. This collision will apply the percentage slow when a character walks within it and then adds the multiplier back when the character exits it. The decal fades but the actor is also destroyed after the duration float * 0.9 since the fade is faster than the duration.

Changes to make in the future: change the colour of the decal based on any status effects tied to this blueprint.