Charge fixes - percentage and movement slow.
Charge was being set back to 0 when finishing the charge resulting in the stats carried over to be 0% of what they should be so this is now fixed by just passing the charge percentage into the correct macro when the spell is fired.

Fixed an issue with the movement speed multiplier not being properly set back to normal or set to the charge speed multiplier in the charge spells ranged attack component.