Movement manager speed multiplier update / fix
Update to movement manager where we now have the speed multiplier and the uncapped speed multiplier. This uncapped speed multiplier is what is used to keep track of the actual speed multiplier and is used to remove and add to the speed multiplier since it is what the speed multiplier actually is and therefore when we change it we are using the uncapped speed multiplier.

What this allows is the calling of the set speed multiplier function with the speed decrease percentage no matter what the speed multiplier is.

Lets say we decrease the speed by 80 percent so the speed multiplier is 0.2 and then whilst this speed decrease is still active another speed decrease happens of 0.7 which would activate the speed multiplier clamp of 0.05 but the actual speed multiplier is -0.5 and so when the first speed decrease stops then the speed multiplier will go to 0.3 since the 70% decrease is still active and then when this stops the speed multiplier will return to 1