Homing functionality.
Homing component functionality implemented.
Activate homing function.
Deactivate homing function.
Update homing target function passing actor parameter for the actor to home on in which is checked to see if it valid.

Begin play we get the spell info and then get the homing info with the following variables set from it: homing strength, homing radius,. auto switch targets and time to switch targets. After this we set the homing acceleration magnitude to be the homing strength and then we check to see if the projectile will automatically switch targets. If it doesn't then the tick is disabled but if it will auto switch then the tick interval is set to be the time to switch targets.

The actual detection for the homing target is the use of a multi sphere trace for pawns with the radius as the homing radius. We only loop through the array if we hit something and we check if the hit actor implements an interface based on who the owner of the projectile is: if it is the player then we check to see if it is the enemy interface and vice versa. Then we check if it is the first index and if so then the homing target is set to the first index in the array. If not we check to see if the current hit actor's distance from the projectile is less than the current homing target's distance from the projectile and if so the current hit actor is set to be the new homing target. After the array is looped through then the activate homing function is called and the update homing target function is called passing the current homing target function.

The only difference in the event tick is calling the deactivate homing function if the trace doesn't hit any pawns.