Custom Collision Presets. Piercing Update.
4 new custom collision presets created in project settings: base spell collision, piercing characters collision, piercing objects collision, and piercing both collision. These are now used in the spell projectile begin play event graph.

Piercing component created. Piercing component is only added to the spell projectile blueprint if the spell has the piercing characteristic.

Had to change around some more data in the piercing struct which made we reload the spell info which made me reload the character info yayayay.

Piercing functionality fully implemented: when the spell's static mesh overlaps with something then it calls the pierced function in the piercing component. Passing the hit thing whether it be the primitive component (for physics bodies), actor for characters and nothing for a hit dynamic object or static object.

Within the pierced function we apply damage where necessary, impulse where necessary for physics bodies and then add onto the number of pierces and if it is the same or greater than the number of max pierces and the max number of pierces variable is true then the spell eol function is called. After piercing then the stats of the spell is changed based on the percentage fields in the piercing data of the spell by setting the spell info directly by casting.