Spells now in character info. Set character info of players in construction script. Spell HUD representation and widgets. Fixed issue with spell names and row names.
Spells are no longer in the weapon structs but instead in the character structs.

Setting the character info from the character name in the character info in the construction script of the third person character - this will later be changed to passing the character info into the player blueprint as it is spawned but need to do the character select screen for this anyway.

Spell names and spell row names weren't matching up so fixed that and that was why the player was having duplicate spells and issues with spells but its fixed now.

New widget blueprint for spells / ability representation.
Setting / updating info function in this widget blueprint.

Creating the HUD widget and adding it to viewport is now in the game mode rather than the level blueprint so can easily access a reference to the HUD in the other blueprints by getting the game mode and casting.

Update spell HUD function in spells component that calls a function in the HUD to then update the specified spell / ability widget appropriately.

Spell / ability widgets set themselves hidden on construction if the passed character index is not greater than or equal to the length of players in the game mode.

When setting the spell / ability name we also split the name based on underscores to separate the words so that they can be automatically wrapped in the text box.