Time to max speed for spells implemented.
Timeline in begin play for lerping the max speed of the projectile which is now also set in the begin play of the projectile to its final max speed which is the spell speed in spell info struct.

We now also set the max speed of the projectile to either be the spell speed or half of the spell speed depending on if the time to max speed is 0 or not.

Also setting the play rate of the timeline before playing it from the beginning by dividing 1 by the time to max speed in the spell struct.