Spell Slow down bool and float for time and slow down start time.
Made a black hole spell. - aoe, charge, slow down.

Spells now have a bool in the general info for spells whether they should slow down and then 2 float variables: one for when the slow down should start and then also how long the slow down should take to reach 0 from the current max speed.
This is done in a similar way to the speed up of the spells.