Attracting spell characteristic info and functionality implemented. Removed print string from health bar. Added in a isvalid check in health bar.
New spell characteristic called attracting which will suck in any nearby characters and physics bodies to the projectile, this new characteristic allows for the creation of the black hole spell.

New struct for attracting characteristic which is in the spell info struct.

Attracting info includes: attracting radius, attracting strength, timed attracting?, attracting duration, delayed attracting?, attracting start time, attracting updating rate.

In the attracting component we get the spell info first and then check if it is a delayed attraction and if so then a delay is done of length attraction start time before setting the tick enabled, if it isn't a delayed attraction then event tick is instantly set enabled. Then we set the component tick interval based on the attracting upating rate, then we check if it is a timed attracting and if so then a delay of length attracting duration is done before setting event tick disabled.

In event tick we constantly do a multisphere trace of radius attracting radius, looking for pawns and physics bodies then we check if we hit anything and if not we clear the force event timer if we did hit something we check if the timer handle is valid and if so we loop through the hits checking if they are in an array of actors and if not we add them to the array, if the timer handle isn't valid then we set the timer by event and set the force event timer variable before doing the looping of the hits.

In the add force custom event we loop through the hit actors and then apply an impulse to characters through their character movement component and we add force to physics bodies hit with the force for physics bodies * 100.