Spell cooldown representation in HUD spell ability widget.
The spell / ability icon now has its color and opacity set within the widget blueprint dependent on the spell being on cooldown. If the spell is on cooldown then the percentage of the cooldown elapsed is the colour and opacity whereas if it isn't on cooldown then the colour and opacity is set to 1 for everything since this is the default.

To do: animation event for when the spell is ready - a little flash or something which will be called after a very small delay of the spell coming off of cooldown.