Got rid of cosmetic error messages from anim blueprint. Added spells to the new character info row that Ben made.
Can check downed info bool added and implemented code to get rid of the cosmetic error messages where the animation blueprints can't access their owner for like a frame at the start of the game.

The reason there were so many errors before with the spells was because when Ben created a new character info row to test with he didn't add any spells to the character so there were countless errors of not having any spells.