Enemy AI implementation.
Pretty janky but the theory is there just needs fixing up and expanding for payload prioritisation and stuff but the targeting of players is there and attacking.

EQS for patrol / passive, melee and ranged attack locations.

Tasks for setting focus target, resetting focus target.

Enemy controller interface.

Added correct interfaces and components to enemy test blueprint.

Multiple trees for passive / idle, chasing player, chasing payload and then main control tree.

Decorators for checking the current state of the enemy controller, and decorator for checking ranges for attacking.

Some changes to the enemy info struct.

Enemy perception and state machine: idle / passive, chasing player and chasing payload that also affect how the perception works.

Will comment all this code in the next couple of days just wanted to power through and get this implemented even if it is janky but this will be fixed and improved.