AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAH. AI fix.
AI now working as intended with targeting the player and switching between melee and ranged combat depending on range to the chasing target and whether they are ranged and can do melee attacks.

Also the payload now has the correct interface on it and can be perceived by enemies now by having the AI stimuli source for sight on it.

Enemies will proritise the payload over the player when the payload is closer and is being perceived by the enemy.

Just need to add in the edge cases where the distance is overridden for setting the target to be the player.