Reviving now uses the heal function instead of directly setting the health. Healing from bloodsplosion based on proximity. Healing from melee attacks based on damage. Enemies do not attack
Healing from bloodsplosion using functionality from Organic Dungeon V2 based on proximity and blood amount.

Also added in functionality for healing from melee attacks added in for the hit detection in the weapon parent blueprint. This is based on the damage dealt divided by 5.

Both healing from bloodsplosion and melee attacks will first check to see if the player is downed or going into being downed and if so the healing won't be applied.

New functionality added in attack task for enemies and new field added in enemies struct for melee attack speed.