Holstered weapon system based on if the character has melee as main weapon in character data table / struct. All functionality within attach weapon component. New socket in skeletal mesh. Changed pivot rotation of staff weapon mesh.
Attach weapon component now has a new function to switch the weapon in your right hand and a new bool for checking if a melee weapon is in your right hand. This function is called within the attack handler for ranged attacks and melee attacks passing the corresponding bool if it is a melee attack.

New field in character info struct on if the character has a melee weapon in their right hand to begin with.

New socket in character mesh on their back for where the weapon is holstered and then also changed the right weapon socket rotation and position a little and then also changed the pivot of the staff static mesh and its sockets.