Reduced damage for players damaging the payload. Friendly fire damage reduced by 50% for players hitting other players. Friendly fire damage reduced for enemies hitting other enemies by 25%. Fixed some payload location things. Fixed another issue with the colour and opacity of spell ability widget. Payload actually takes damage from all spell related damage sources now.
Damage taken by payload in take damage function in payload health component is now multiplied by 0.5 so payload will only take 50% damage from players.

This is similar now in the actual health component where the damage taken is multiplied by 0.5 if both the character dealing damage and the character taking damage both implement the player interface. Then if both the character taking damage and the character dealing damage implement the enemy interface then the damage is multiplied by 0.75 so enemies will take 75% of the damage being dealt if another enemy is damaging them.

In the payload health component set the epicentre of the camera shake correctly and the play sound at location correctly.

In the spell / ability widget if the spell cooldown index is not valid then the colour and opacity is set to be 1,1,1,1 so that the colour and opacity doesn't carry over and it gives the illusion that each spell / ability is individually having its colour and opacity changed.

Added in checks for if the hit actor implements the payload interface anyway where spell related things were dealing damage (calling the take damage function) and will now call the take damage function in the payload health component if the hit actor is the payload.