Out of combat healing implemented. In combat checker component created and functionality implemented. Fixed owner dealing damage from melee attacks in weapon blueprint. When downed the hitbox is now correct and is reset properly on revive.
New component created: in combat checker which has a custom event that sets in combat true and then starts a re-triggerable delay of a variable amount of time set with the variable out of combat time, which after the delay sets in combat false. Then in event tick we check if the health component is valid and if so then we check if we are not in combat and not downed or in the downed transition - this may need to be updated to check if not in reviving transition if this is how reviving will be implemented - if the health component is not valid then we only check if we are not in combat. If we are in combat then we check if the heal timer is true and if so we invalidate and clear it but if we are not in combat and the heal timer is not valid then we set a timer by the heal event of time heal interval ( a float variable in the component) and set the heal timer.

In the heal custom event we check if the owner is the payload and if so we call the new heal function in the payload health component with the amount to heal the heal amount payload variable in this component and if the owner isn't the payload we check if the owner is a player and if so we call the heal function in the health component with the amount to heal the heal amount player variable in this component.

The take damage function calls in the weapon parent for melee hits wasn't correct with passing who was dealing the damage since it needs to be a character / player / pawn not the weapon itself so now it is passing the owner of the weapon which means that the friendly fire reduced damage now works properly for melee attacks.

On death going downed we now set the half height of the capsule component of the character to be 35 instead of 90 and also set the skeletal mesh local position to be 0 instead of -89. Then when reviving we set these back to their defaults. Also to make this work I had to save the reviving player ref as a new variable after the reviving checks are found since it was being made invalid during the process and wasn't allowing for changes properly. This might need to be changed in the future when reviving is further explored and fully implemented.