LOTS OF UI. Game instance created as well in instances folder.
Made screens for all the different screens needed.
Also created a clickable button widget with various functions within it that will be used in all screens. Also created some enums for the clickable button functions, button hovered functions and return to screens enum.
These control what the button will do when it is clicked, what the button will do when it is hovered and then what screen will be returned to when it is clicked which is used for return buttons on screens however this is not fully implemented yet as I haven't made all the screens and only really made the main menu screen.

Look at clickable button functions and custom events for more info.

Also created a new game mode for the main menu which stores references to various widget screens to check if they are already valid and to store their references. Then when the pause menu is fully implemented then the thirdpersongamemode will be used to store a reference to the pause menu screen widget.