Controller support / focus functionality for the clickable buttons implemented YAYAYAYYAYAY.
We have to set the focus to be one of the buttons first in each screen so remember to do that in each screen. The start of the main menu screen is done in the level blueprint for setting the focus.

Then we make sure everything that needs to be focusable is focusable.

Then in the clicked button widget or anything else that needs focus we use the on added to focus path event for when it is in focus like when it is hovered by the mouse and then  for when it is no longer in focus we use the on removed from focus path event.

Obviously when more co-op functionality is added for menus will need to set the focus for each player controller not just the first one.