Button saturate on hover so more noticeable. Main Menu game mode update. Multi-controller testing for character select. New blueprint that will work for showing the players in the character select menu.
Buttons now also saturate more when they are hovered so that it is more noticeable when they are hovered.

Updated the main menu game mode to be able to spawn the new player blueprint for the character select which is called when the character select menu is constructed.

Multiple controllers tested for the character select menu which works as intended with allowing multiple players to select different buttons at once.

New blueprint for showing the characters in the character select menu that they will be able to rotate. Copied the thirdpersoncharacter blueprint and removed a lot of unnecessary functionality. Will need to add functionality for updating the info of these blueprints and therefore also updating the meshes of the blueprints and showing the weapons attached so will also need to add the weapon attach component to this blueprint.

Need to finally add the functionality for "joining" the character select menu which will be when a interact button is pressed as we will create a new IMC for the character menu. Then need to do some more functionality on readying up and then also starting the game and transmitting the data to the game instance and then into the spawned characters through the thirdpersongamemode.